feat: Configure cluster issuer for DNS (#54)
This configures the Let's Encrypt cluster issuer to follow CNAME records
in order to determine the right zone for DNS challenges.

With this, we can issue certificates for the zone
'freifunk-duesseldorf.de' for certain hostnames without giving
cert-manager permission to edit this zone.

see-also: https://github.com/ffddorf/terraform-cloudflare-dns/pull/48